Update variables.md explanation of parameters
I was always confused by this example at multiple levels. I didn't understand "constrained on creation to a Parameter set" because I didn't understand what that set is, is there a whole set of Parameters? I was unsure what `Parameter(i)` means, because `i` looks like an index, not a value. I also didn't get how there can be `x in SecondOrderCone` with same syntax as `x in Parameter`, they felt like different concepts. 

It has finally dawned on me that `x in S` always means constrained within a set `S`, and `Parameter(1)` is just a very small set containing the number (not the index 1). Part of the confusion is that the values of `p[i]` are only revealed farther below the example and not immediately.

I'm proposing different wording to specifically address my obtuseness. I hope this, or something better, can clarify this confusion.